requirements.txt
mysql-connector-python
sqlalchemy
fastapi
uvicorn
python-dotenv
yt-dlp
ffmpeg-python
pytesseract
openai-whisper
pydub
pyaudio
boto3
numpy
pytesseract
Pillow
ImageHash
googlemaps
openai
python-decouple

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1

# Environment variables
python-dotenv>=0.19.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Utilities
requests>=2.26.0
pydantic>=1.8.2

# Maps
folium>=0.12.1

# CloudWatch Logging
watchtower>=3.0.0
boto3>=1.26.0